Add difficulty for quenching

+++ Assets/Scripts/MiniGames/AMiniGameManager.cs
@@ -5,7 +5,7 @@ namespace Minigame
     public delegate void MinigameCallBack();
     public abstract class AMiniGameManager : MonoBehaviour
     {
-        public virtual void RunMinigame(MinigameCallBack cb_result, int difficultyFactor)
+        public virtual void RunMinigame(MinigameCallBack cb_result, float difficultyFactor)
         {
             gameObject.SetActive(true);
         }
